// src/types/spaces.ts
export interface AuthenticatePeriscopeResponse {
  data: {
    authenticate_periscope: string;
  };
  errors?: any[];
}

export interface LoginTwitterTokenResponse {
  cookie: string;
  user: {
    class_name: string;
    id: string;
    created_at: string;
    is_beta_user: boolean;
    is_employee: boolean;
    is_twitter_verified: boolean;
    verified_type: number;
    is_bluebird_user: boolean;
    twitter_screen_name: string;
    username: string;
    display_name: string;
    description: string;
    profile_image_urls: {
      url: string;
      ssl_url: string;
      width: number;
      height: number;
    }[];
    twitter_id: string;
    initials: string;
    n_followers: number;
    n_following: number;
  };
  type: string;
}